Tidied up code

+++ UWPChip8/Chip8.Core/CPU.Instructions.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Chip8.Core
 {
@@ -383,7 +379,7 @@ namespace Chip8.Core
         void DRW(byte ax, byte ay, byte nibble)
         {
             _v[0xf] = 0;
-            byte sprite = 0x0;
+            int sprite = 0x0;
             int start = I.GetValue();
 
             for (int y = 0; y < nibble; y++)
@@ -469,17 +465,7 @@ namespace Chip8.Core
         /// <param name="ax">ax</param>
         void BCD(byte ax)
         {
-            // This appears to be used primarily for scoreboards, seeking correction?
-
-            // Example Decimal Representation
-            // _v[ax] = decimal 10
-            // Results (3) = 10, 1, 0
-
-            // Therefore I = 10, I + 1 = 1, I + 2 = 0;
-
-            // Seems that in pong at least the values are stored at index 754. Incorrectly the scores
-            // are always equal to each other, ie. 4 - 4, when 4 points are scored on, at least,
-            // left hand side and no points on right hand side. Slight issue here :)
+            // This appears to be used primarily for scoreboards
             var number = _v[ax];
 
             for (int i = 3; i > 0; i--)
